Slowing down arm
